MidiHead Version 0.0.1
Y'know, it's a shame that more people don't check my github. This is the only place where i'm going to let everyone know that my full project release is the 26th of September 2020, available on Youtube and most music platforms. You can comb through the code and get a pretty quick idea of what is going on. There probably won't be a Version 0.0.2 - this works great.